Restructure utilities to be first-class exports
