+++ packages/graphql-mocks/src/graphql/index.ts
@@ -1,1 +1,3 @@
 export { GraphQLHandler } from './handler';
+export * as utils from './utils';
+export * as typeUtils from './type-utils';
