+++ packages/graphql-mocks/src/graphql/type-utils/has-list-type.ts
@@ -1,0 +1,8 @@
+import { GraphQLType, isListType, isNonNullType } from 'graphql';
+
+/**
+ * Checks if a type is a list type or a wrapped list type (ie: wrapped with non-null)
+ */
+export function hasListType(type: GraphQLType): boolean {
+  return isListType(type) || (isNonNullType(type) && isListType(type.ofType));
+}
