+++ packages/graphql-mocks/src/graphql/type-utils/is-internal-type.ts
@@ -1,0 +1,13 @@
+import { GraphQLType, isNamedType } from 'graphql';
+
+export function isInternalType(type: GraphQLType | string): boolean {
+  if (isNamedType(type)) {
+    type = type.name;
+  }
+
+  if (typeof type !== 'string') {
+    return false;
+  }
+
+  return type.startsWith('__');
+}
