+++ packages/graphql-mocks/src/graphql/utils/attach-type-resolvers-to-schema.ts
@@ -1,0 +1,19 @@
+import { GraphQLSchema, isAbstractType } from 'graphql';
+import { ResolverMap } from '../../types';
+
+export function attachTypeResolversToSchema(schema: GraphQLSchema, resolverMap: ResolverMap): void {
+  for (const typeName in resolverMap) {
+    const type = schema.getType(typeName);
+
+    //  Note: __resolveType for type resolvers is a convention borrowed from
+    //  graphql-tools resolver maps. This allows a single ResolverMap to be used
+    // for both type resolvers for abstract types (unions & interfaces), as well
+    // as field resolvers
+    const typeResolver = resolverMap[typeName].__resolveType;
+    const hasTypeResolver = Boolean(typeResolver);
+
+    if (hasTypeResolver && isAbstractType(type)) {
+      type.resolveType = typeResolver;
+    }
+  }
+}
