+++ packages/graphql-mocks/src/graphql/utils/copy-schema.ts
@@ -1,0 +1,5 @@
+import { GraphQLSchema, buildSchema, printSchema } from 'graphql';
+
+export function copySchema(schema: GraphQLSchema): GraphQLSchema {
+  return buildSchema(printSchema(schema));
+}
