+++ packages/graphql-mocks/src/highlight/index.ts
@@ -9,3 +9,4 @@ export { type } from './highlighter/type';
 export { union } from './highlighter/union';
 
 export { HIGHLIGHT_ALL, HIGHLIGHT_ROOT_QUERY, HIGHLIGHT_ROOT_MUTATION } from './highlighter/constants';
+export * as utils from './utils';
