+++ packages/graphql-mocks/src/pack/utils/embed-pack-options-in-context.ts
@@ -1,0 +1,12 @@
+import { ResolverContext } from '../../types';
+import { PackOptions } from '../types';
+
+export function embedPackOptionsInContext(context: Record<string, unknown>, packOptions: PackOptions): ResolverContext {
+  context = context ?? {};
+  context = {
+    ...context,
+    pack: context.pack || packOptions,
+  };
+
+  return context;
+}
