+++ packages/graphql-mocks/src/pack/utils/index.ts
@@ -1,0 +1,1 @@
+export { embedPackOptionsWrapper } from './embed-pack-options-wrapper';
