+++ packages/graphql-mocks/src/pack/utils/normalize-pack-options.ts
@@ -1,0 +1,14 @@
+import { PackOptions } from '../types';
+
+export const defaultPackOptions: PackOptions = { state: {}, dependencies: {} };
+
+export function normalizePackOptions(packOptions: Partial<PackOptions> = defaultPackOptions): PackOptions {
+  const normalized = {
+    ...defaultPackOptions,
+
+    dependencies: packOptions.dependencies ?? defaultPackOptions.dependencies,
+    state: packOptions.state ?? defaultPackOptions.state,
+  };
+
+  return normalized;
+}
