+++ packages/graphql-mocks/src/relay/create-edge.ts
@@ -1,0 +1,6 @@
+export function createEdge<T>(node: T, cursor: string): { cursor: string; node: T } {
+  return {
+    cursor,
+    node,
+  };
+}
