+++ packages/graphql-mocks/src/relay/index.ts
@@ -1,1 +1,4 @@
-export { applyCursorsToEdges, relayPaginateNodes } from './utils';
+export { isRelayConnectionField } from './is-relay-connection-field';
+export { applyCursorsToEdges } from './apply-cursor-to-edges';
+export { createEdge } from './create-edge';
+export { paginateNodes } from './paginate-nodes';
