+++ packages/graphql-mocks/src/resolver-map/index.ts
@@ -2,3 +2,4 @@ export { embed } from './embed';
 export { layer } from './layer';
 export { setResolver } from './set-resolver';
 export { getResolver } from './get-resolver';
+export * as utils from './utils';
