+++ packages/graphql-mocks/src/resolver-map/utils/index.ts
@@ -1,0 +1,1 @@
+export { highlightAllCallback } from './highlight-all-callback';
