+++ packages/graphql-mocks/src/resolver/utils/is-field-resolver.ts
@@ -1,0 +1,6 @@
+import { GraphQLType, isObjectType } from 'graphql';
+import { FieldResolver } from '../../types';
+
+export function isFieldResolver(type: GraphQLType, resolver: unknown): resolver is FieldResolver {
+  return Boolean(isObjectType(type) && resolver);
+}
